# movil/openapi_v2.yaml
openapi: 3.0.3
info:
  title: Escandio API movil v2
  version: 2.0.0
  description: Contrato de la API v2 consumida por la app movil berkelio. Ver movil/contrato_v2.py.
paths:
...
  /api/v2/seguimiento/{id}/responder/:
    post:
      operationId: seguimiento_responder_create
      description: Consultas de seguimiento dirigidas al conductor.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este rut seguimiento.
        required: true
      tags:
      - seguimiento
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponderConsultaRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ResponderConsultaRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ResponderConsultaRequestRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Id'
          description: ''
  /api/v2/seguimiento/mensaje/:
    post:
      operationId: seguimiento_mensaje_create
      description: Consultas de seguimiento dirigidas al conductor.
      tags:
      - seguimiento
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/IdRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IdRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Id'
          description: ''
  /api/v2/seguimiento/mi-conversacion/:
    get:
      operationId: seguimiento_mi_conversacion_retrieve
      description: Consultas de seguimiento dirigidas al conductor.
      tags:
      - seguimiento
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultaPendienteMovil'
          description: ''
  /api/v2/seguimiento/pendientes/:
    get:
      operationId: seguimiento_pendientes_list
      description: Consultas de seguimiento dirigidas al conductor.
      parameters:
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      tags:
      - seguimiento
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedConsultaPendienteMovilList'
          description: ''
...
components:
  schemas:
...
    ConsultaPendienteMovil:
      type: object
      description: Consulta '¿como va el viaje?' que el conductor tiene por responder.
      properties:
        id:
          type: integer
          readOnly: true
        despacho_id:
          type: integer
          nullable: true
          readOnly: true
        pregunta:
          type: string
        opciones:
          readOnly: true
          nullable: true
        fecha_registro:
          type: string
          format: date-time
          readOnly: true
      required:
      - despacho_id
      - fecha_registro
      - id
      - opciones
      - pregunta
...
    Id:
      type: object
      properties:
        id:
          type: integer
      required:
      - id
    IdRequest:
      type: object
      properties:
        id:
          type: integer
      required:
      - id
...
    PaginatedConsultaPendienteMovilList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ConsultaPendienteMovil'
...
    ResponderConsultaRequestRequest:
      type: object
      description: Documenta el cuerpo de POST /seguimiento/<id>/responder/.
      properties:
        opcion:
          type: string
          minLength: 1
          description: Opcion elegida de las ofrecidas.
        comentario:
          type: string
          nullable: true
        movil_token:
          type: string
          minLength: 1
          description: Clave de idempotencia generada por la app.
        fecha:
          type: string
          minLength: 1
          description: 'Formato: YYYY-MM-DD HH:MM'
      required:
      - fecha
      - movil_token
      - opcion
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT